simplify to types
