--- consensus/src/threshold_simplex/wire.proto
@@ -1,78 +1,0 @@
-syntax = "proto3";
-package threshold_simplex.wire;
-
-message Voter {
-    oneof payload {
-        // To avoid complex resolution scenarios, we include partial signatures/signatures for both
-        // consensus and the seed in the same message (dropping anything that doesn't have both).
-        Notarize notarize = 1;
-        Notarization notarization = 2;
-        Nullify nullify = 3;
-        Nullification nullification = 4;
-        Finalize finalize = 5;
-        Finalization finalization = 6;
-    }
-}
-
-// To ensure all participants stay informed, Proposal is included in all Notarize/Finalize messages.
-message Proposal {
-    uint64 view = 1;
-    uint64 parent = 2;
-    bytes payload = 3; // Digest
-}
-
-// A Notarize from the leader for a view is also treated as a proposal
-// for that view.
-message Notarize {
-    Proposal proposal = 1;
-    bytes proposal_signature = 2; // Partial
-    bytes seed_signature = 3; // Partial
-}
-
-message Notarization {
-    Proposal proposal = 1;
-    bytes proposal_signature = 2;
-    bytes seed_signature = 3;
-}
-
-// Nullify is used to indicate that a view should be nullified (skipped).
-message Nullify {
-    uint64 view = 1;
-    bytes view_signature = 2; // Partial
-    bytes seed_signature = 3; // Partial
-}
-
-message Nullification {
-    uint64 view = 1;
-    bytes view_signature = 2;
-    bytes seed_signature = 3;
-}
-
-message Finalize {
-    Proposal proposal = 1;
-    bytes proposal_signature = 2; // Partial
-}
-
-message Finalization {
-    Proposal proposal = 1;
-    bytes proposal_signature = 2;
-    bytes seed_signature = 3;
-}
-
-message Backfiller {
-    uint64 id = 1;
-    oneof payload {
-        Request request = 2;
-        Response response = 3;
-    }
-}
-
-message Request {
-    repeated uint64 notarizations = 1;
-    repeated uint64 nullifications = 2;
-}
-
-message Response {
-    repeated Notarization notarizations = 1;
-    repeated Nullification nullifications = 2;
-}
